restructure 'render' module
Split modules out into files to keep individual module size more
managable.
